# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bipupu",
    "description": "BIPUPU API 服务",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/public/register": {
      "post": {
        "tags": [
          "认证",
          "认证"
        ],
        "summary": "Register User",
        "description": "用户注册\n\n参数：\n- username: 用户名，用于登录\n- password: 密码，长度限制为6-128字符\n- nickname: 昵称，可选\n\n返回：\n- 成功：返回新创建的用户信息\n- 失败：400（验证失败）或429（注册频率限制）或500（数据库操作失败）\n\n特性：\n- 速率限制：每个用户名每小时最多10次注册尝试\n- 密码自动哈希存储\n- 自动生成唯一的 bipupu_id\n- 验证用户名和密码格式\n\n注意：\n- 用户名必须是唯一的\n- 密码在传输和存储中都会加密\n- 注册后用户自动处于活跃状态",
        "operationId": "register_user_api_public_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/login": {
      "post": {
        "tags": [
          "认证",
          "认证"
        ],
        "summary": "Login",
        "description": "用户登录\n\n参数：\n- username: 用户名\n- password: 密码\n\n返回：\n- 成功：返回访问令牌和刷新令牌\n- 失败：401（用户名或密码错误）或400（用户未激活）或429（登录频率限制）\n\n包含信息：\n- access_token: 访问令牌，用于API认证\n- refresh_token: 刷新令牌，用于获取新的访问令牌\n- token_type: 令牌类型，固定为\"bearer\"\n- expires_in: 访问令牌过期时间（秒）\n- user: 当前用户基本信息\n\n特性：\n- 速率限制：每个用户名每分钟最多5次登录尝试\n- 支持JWT令牌认证\n- 自动验证用户激活状态\n\n注意：\n- 访问令牌有过期时间，需定期刷新\n- 刷新令牌用于获取新的访问令牌\n- 非活跃用户无法登录",
        "operationId": "login_api_public_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/refresh": {
      "post": {
        "tags": [
          "认证",
          "认证"
        ],
        "summary": "Refresh Token",
        "description": "刷新访问令牌\n\n参数：\n- refresh_token: 刷新令牌\n\n返回：\n- 成功：返回新的访问令牌和刷新令牌\n- 失败：401（无效的刷新令牌）或429（刷新频率限制）\n\n包含信息：\n- access_token: 新的访问令牌\n- refresh_token: 新的刷新令牌（支持令牌轮换）\n- token_type: 令牌类型，固定为\"bearer\"\n- expires_in: 访问令牌过期时间（秒）\n\n特性：\n- 速率限制：每个刷新令牌每小时最多10次刷新尝试\n- 支持令牌轮换，每次刷新生成新的刷新令牌\n- 验证刷新令牌类型和用户状态\n- 自动验证用户是否活跃\n\n注意：\n- 刷新令牌只能用于刷新访问令牌，不能用于API访问\n- 旧的刷新令牌在刷新后失效\n- 非活跃用户的刷新令牌无效",
        "operationId": "refresh_token_api_public_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRefresh"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/logout": {
      "post": {
        "tags": [
          "认证",
          "认证"
        ],
        "summary": "Logout",
        "description": "用户登出\n\n参数：\n- Bearer Token: 通过Authorization头传递的访问令牌\n\n返回：\n- 成功：返回登出成功消息\n- 失败：401（无效令牌）或500（服务器错误）\n\n特性：\n- 将当前访问令牌加入黑名单\n- 黑名单令牌在剩余有效期内无法使用\n- 支持令牌过期时间计算\n\n注意：\n- 需要有效的访问令牌\n- 登出后当前令牌立即失效\n- 其他设备的令牌仍然有效，需要分别登出\n- 刷新令牌不受登出影响，需要单独处理",
        "operationId": "logout_api_public_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CosmicProfile": {
        "properties": {
          "birthday": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthday",
            "description": "公历生日，格式 YYYY-MM-DD"
          },
          "zodiac": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zodiac",
            "description": "西方星座，例如 '白羊座'"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age",
            "description": "年龄（整数岁）"
          },
          "bazi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bazi",
            "description": "生辰八字字符串（可由用户提供或第三方服务计算）"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender",
            "description": "性别，例如 'male','female','other'，可选"
          },
          "mbti": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mbti",
            "description": "MBTI 类型，例如 'INTJ'"
          },
          "birth_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Time",
            "description": "出生时间，例如 '08:30'"
          },
          "birthplace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthplace",
            "description": "出生地，城市/经纬度等"
          }
        },
        "type": "object",
        "title": "CosmicProfile",
        "description": "结构化的宇宙信息（可在未来扩展）"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StatusResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "StatusResponse",
        "description": "通用状态响应"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "Token",
        "description": "令牌响应模式"
      },
      "TokenRefresh": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "TokenRefresh",
        "description": "刷新令牌模式"
      },
      "UserCreate": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserCreate",
        "description": "创建用户模式"
      },
      "UserLogin": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserLogin",
        "description": "用户登录模式"
      },
...
      "UserResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "bipupu_id": {
            "type": "string",
            "title": "Bipupu Id"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "cosmic_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CosmicProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser",
            "default": false
          },
          "last_active": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Active",
            "description": "最后活跃时间"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "username",
          "bipupu_id",
          "id",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "用户响应模式"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}